Uploaded .py file
This is the .py to place in the /apps folder of your Cardputer running MicroHydra.  This should work when running in Thonny connected to your Cardputer via USB.  Known issue: when scrolling airports in Select Airport, the screen does not scroll, so you can only see the first few options.  Implementing scrolling in this mode is on the to do list.