Large update adding SSID connection management
This update allows you to view available wifi networks and enter the password to connect.  The password will be stored on the Pico in a .txt file, so of course it will be important to know this when considering your network's security, as physical access to the Pico will allow reading of the text file with Thonny etc.